GNOME 45: Fix-up extension.js
GNOME 45 requires extension.js to export a default class that
extends Extension (from the module extensions/extension.js)
and has two methods: enable() and disable(). The rest of the
changes are to adapt the existing code to this change.
NOTE: imports.misc.extensionUtils.getCurrentExtension() does not
exist anymore, and I couldn't get Extension.lookupByUUID() and
Extension.lookupByURL() to work, even though they're supposed to
be the replacement.
NOTE: It's not necessary to load the settings schema manually,
GNOME does it automatically with ExtensionBase.getSettings()