GNOME 45: Fix-up prefs.js
GNOME 45 requires prefs.js to export a default class that extends
ExtensionPreferences (from the module extensions/prefs.js) and
has the method fillPreferencesWindow(). The rest of the changes
have been made to adapt the existing code to this change.
NOTE: Api.trigger_oauth() now takes an argument, that being the
extension path. This was done because I couldn't find a way to
get the extension object in the api.js module.
NOTE: It's not necessary to load the settings schema manually,
GNOME does it automatically with ExtensionBase.getSettings()