Port to GNOME 45 (#101)
* GNOME 45: Update metadata.json

* GNOME 45: Fix-up import statements

* GNOME 45: Replace ByteArray.toString() with TextDecoder.decode()

The ByteArray.toString() method is deprecated and not available
in GNOME 45. The documentation suggests to use
TextDecoder.decode() instead, so that what's been done in
this commit.

* GNOME 45: Fix-up extension.js

GNOME 45 requires extension.js to export a default class that
extends Extension (from the module extensions/extension.js)
and has two methods: enable() and disable(). The rest of the
changes are to adapt the existing code to this change.
NOTE: imports.misc.extensionUtils.getCurrentExtension() does not
exist anymore, and I couldn't get Extension.lookupByUUID() and
Extension.lookupByURL() to work, even though they're supposed to
be the replacement.
NOTE: It's not necessary to load the settings schema manually,
GNOME does it automatically with ExtensionBase.getSettings()

* GNOME 45: Fix-up prefs.js

GNOME 45 requires prefs.js to export a default class that extends
ExtensionPreferences (from the module extensions/prefs.js) and
has the method fillPreferencesWindow(). The rest of the changes
have been made to adapt the existing code to this change.
NOTE: Api.trigger_oauth() now takes an argument, that being the
extension path. This was done because I couldn't find a way to
get the extension object in the api.js module.
NOTE: It's not necessary to load the settings schema manually,
GNOME does it automatically with ExtensionBase.getSettings()

* GNOME 45: Export modules' functions